refactor : console command description message 기능 수정
