+++ src/main/java/org/programmers/kdt/weekly/command/io/InputErrorType.java
@@ -1,10 +1,11 @@
-package org.programmers.kdt.weekly.io;
+package org.programmers.kdt.weekly.command.io;
 
 public enum InputErrorType {
-    BLACK_LIST_EMPTY("There are no saved blacklist."),
+    CUSTOMER_EMPTY("There are no saved customers."),
     COMMAND("This command does not exist."),
     INVALID("Invalid input."),
-    VOUCHER_EMPTY("There are no saved vouchers.");
+    VOUCHER_EMPTY("There are no saved vouchers."), VOUCHER_WALLET_EMPTY("There are no saved Voucher Wallet."),
+    DUPLICATE_EMAIL("This email already exists.");
 
     private final String message;
 
@@ -15,4 +16,5 @@ public enum InputErrorType {
     public static String getMessage(InputErrorType inputErrorType) {
         return inputErrorType.message;
     }
+
 }
